refactor(popup): 開閉アニメーション制御をhookに切り出す
