+++ packages/wiz-ui-react/src/components/base/popup/utils/index.ts
@@ -1,2 +1,1 @@
 export * from "./placement";
-export * from "./fade-animation";
